requirements.txt
python-dotenv
llama-index
llama-parse
llama-index-multi-modal-llms-openai
llama-index-vector-stores-chroma
llama-index-postprocessor-cohere-rerank
llama-index-retrievers-bm25
chromadb
requests
openai
streamlit